+++ lib/numpy.dart
@@ -1,3 +1,4 @@
+export 'src/expr/expr.dart';
 export 'src/numpy/linalg.dart';
 export 'src/numpy/numpy.dart';
 export 'src/numpy/random.dart';